Memoise useSafeIntl
previously it was not memoized, so when we used it as a dep for useMemo and useCallback
these were always called.

e.g. In the column calculation on the column page, when the dialog was opened, the useMemo for the column changed
and the whole table was compaed and it took two seconds to display the dialog, leaving a hanging effect.
It now takes 0.5s to display the dialog which is still on the long side but less perceptible by the user